@@ -13,7 +13,6 @@ import com.dropbox.core.v2.files.Metadata;
 
 import java.io.File;
 import java.util.ArrayList;
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
 
@@ -21,6 +20,7 @@ import fr.ralala.worktime.MainApplication;
 import fr.ralala.worktime.R;
 import fr.ralala.worktime.sql.SqlHelper;
 import fr.ralala.worktime.utils.AndroidHelper;
+import fr.ralala.worktime.ui.utils.UIHelper;
 
 /**
  *******************************************************************************
@@ -33,16 +33,16 @@ import fr.ralala.worktime.utils.AndroidHelper;
  */
 public class DropboxImportExport implements DropboxListener{
   private static final String PATH = "";
-  private DropboxHelper helper = null;
-  private ProgressDialog dialog = null;
-  private File file = null;
-  private Context c = null;
-  private DropboxUploaded dropboxUploaded = null;
-  private DropboxDownloaded dropboxDownloaded = null;
+  private DropboxHelper mHelper = null;
+  private ProgressDialog mDialog = null;
+  private File mFile = null;
+  private Context mContext = null;
+  private DropboxUploaded mDropboxUploaded = null;
+  private DropboxDownloaded mDropboxDownloaded = null;
 
 
   public DropboxImportExport() {
-    helper = DropboxHelper.helper();
+    mHelper = DropboxHelper.helper();
   }
 
   public interface DropboxUploaded {
@@ -52,41 +52,35 @@ public class DropboxImportExport implements DropboxListener{
     void dropboxDownloaded(final boolean error);
   }
 
-  public boolean importDatabase(final Context c, final DropboxDownloaded dropboxDownloaded) {
-    this.c = c;
-    this.dropboxDownloaded = dropboxDownloaded;
-    if(helper.connect(c, c.getString(R.string.app_key))) {
-      if(dialog == null)
-        dialog = buildProgress(c);
-      dialog.show();
-      new ListFolderTask(helper.getClient(), this).execute(PATH);
-      return true;
+  public void importDatabase(final Context c) {
+    mContext = c;
+    mDropboxDownloaded = null;
+    if(mHelper.connect(c, c.getString(R.string.app_key))) {
+      if(mDialog == null)
+        mDialog = buildProgress(c);
+      mDialog.show();
+      new ListFolderTask(mHelper.getClient(), this).execute(PATH);
     }
-    return false;
-  }
-
-  public boolean exportDatabase(final Context c, final DropboxUploaded dropboxUploaded) {
-    return exportDatabase(c, true, dropboxUploaded);
   }
 
   public boolean exportDatabase(final Context c, boolean displayDialog, final DropboxUploaded dropboxUploaded) {
-    this.c = c;
-    this.dropboxUploaded = dropboxUploaded;
-    if(dialog == null && displayDialog)
-      dialog = buildProgress(c);
-    if(helper.connect(c, c.getString(R.string.app_key))) {
-      if(dialog != null)
-        dialog.show();
+    mContext = c;
+    mDropboxUploaded = dropboxUploaded;
+    if(mDialog == null && displayDialog)
+      mDialog = buildProgress(c);
+    if(mHelper.connect(c, c.getString(R.string.app_key))) {
+      if(mDialog != null)
+        mDialog.show();
       try {
         File path = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
-        file = new File(SqlHelper.copyDatabase(c, SqlHelper.DB_NAME, path.getAbsolutePath()));
-        new UploadFileTask(c, helper.getClient(), this).execute(Uri.fromFile(file).toString(), PATH);
+        mFile = new File(SqlHelper.copyDatabase(c, SqlHelper.DB_NAME, path.getAbsolutePath()));
+        new UploadFileTask(c, mHelper.getClient(), this).execute(Uri.fromFile(mFile).toString(), PATH);
         return true;
       } catch(Exception e) {
         safeRemove();
-        if(dialog != null)
-          dialog.dismiss();
-        AndroidHelper.toast_long(c, c.getString(R.string.error) + ": " + e.getMessage());
+        if(mDialog != null)
+          mDialog.dismiss();
+        UIHelper.toast_long(c, c.getString(R.string.error) + ": " + e.getMessage());
         Log.e(getClass().getSimpleName(), "Error: " + e.getMessage(), e);
       }
     }
@@ -94,9 +88,10 @@ public class DropboxImportExport implements DropboxListener{
   }
 
   private void safeRemove() {
-    if(file != null) {
-      file.delete();
-      file = null;
+    if(mFile != null) {
+      //noinspection ResultOfMethodCallIgnored
+      mFile.delete();
+      mFile = null;
     }
   }
 
@@ -110,75 +105,70 @@ public class DropboxImportExport implements DropboxListener{
 
   @Override
   public void onDropboxUploadComplete(FileMetadata result) {
-    if(dialog != null)
-      dialog.dismiss();
-    AndroidHelper.toast(c, c.getString(R.string.export_success));
+    if(mDialog != null)
+      mDialog.dismiss();
+    UIHelper.toast(mContext, mContext.getString(R.string.export_success));
     safeRemove();
-    if(dropboxUploaded != null)
-      dropboxUploaded.dropboxUploaded(false);
+    if(mDropboxUploaded != null)
+      mDropboxUploaded.dropboxUploaded(false);
   }
 
   @Override
   public void onDropboxUploadError(Exception e) {
-    if(dialog != null)
-      dialog.dismiss();
+    if(mDialog != null)
+      mDialog.dismiss();
     Log.e(getClass().getSimpleName(), "Failed to upload file.", e);
-    AndroidHelper.toast(c, c.getString(R.string.error_dropbox_upload));
+    UIHelper.toast(mContext, mContext.getString(R.string.error_dropbox_upload));
     safeRemove();
-    if(dropboxUploaded != null)
-      dropboxUploaded.dropboxUploaded(true);
+    if(mDropboxUploaded != null)
+      mDropboxUploaded.dropboxUploaded(true);
   }
 
   @Override
   public void onDroptboxDownloadComplete(File result) {
-    if(dialog != null)
-      dialog.dismiss();
+    if(mDialog != null)
+      mDialog.dismiss();
     try {
-      loadDb(c, result);
+      loadDb(mContext, result);
     } catch(Exception e) {
-      AndroidHelper.toast_long(c, c.getString(R.string.error) + ": " + e.getMessage());
+      UIHelper.toast_long(mContext, mContext.getString(R.string.error) + ": " + e.getMessage());
       Log.e(getClass().getSimpleName(), "Error: " + e.getMessage(), e);
     }
     safeRemove();
-    if(dropboxDownloaded != null)
-      dropboxDownloaded.dropboxDownloaded(false);
+    if(mDropboxDownloaded != null)
+      mDropboxDownloaded.dropboxDownloaded(false);
   }
 
   @Override
   public void onDroptboxDownloadError(Exception e) {
-    if(dialog != null)
-      dialog.dismiss();
+    if(mDialog != null)
+      mDialog.dismiss();
     Log.e(getClass().getSimpleName(), "Failed to download file.", e);
-    AndroidHelper.toast(c, c.getString(R.string.error_dropbox_download));
+    UIHelper.toast(mContext, mContext.getString(R.string.error_dropbox_download));
     safeRemove();
-    if(dropboxDownloaded != null)
-      dropboxDownloaded.dropboxDownloaded(true);
+    if(mDropboxDownloaded != null)
+      mDropboxDownloaded.dropboxDownloaded(true);
   }
 
 
   @Override
   public void onDroptboxListFoderDataLoaded(ListFolderResult result) {
-    if(dialog != null)
-      dialog.dismiss();
+    if(mDialog != null)
+      mDialog.dismiss();
     List<Metadata> list = result.getEntries();
-    Collections.sort(list, new Comparator<Metadata>() {
-      @Override
-      public int compare(Metadata m1, Metadata m2) {
-        return m1.getName().compareTo(m2.getName());
-      }
-    });
+    list.sort(Comparator.comparing(Metadata::getName));
     if (list.isEmpty())
-      AndroidHelper.toast_long(c, c.getString(R.string.error_no_files));
+      UIHelper.toast_long(mContext, mContext.getString(R.string.error_no_files));
     else {
-      computeAndLoad(c, list, new AndroidHelper.AlertDialogListListener<Metadata>() {
+      computeAndLoad(mContext, list, new UIHelper.AlertDialogListListener<Metadata>() {
         @Override
         public void onClick(final Metadata m) {
           try {
-            if(dialog != null)
-              dialog.show();
-            new DownloadFileTask(c, helper.getClient(), DropboxImportExport.this).execute((FileMetadata)m);
+            if(mDialog != null)
+              mDialog.show();
+            new DownloadFileTask(mContext, mHelper.getClient(), DropboxImportExport.this).execute((FileMetadata)m);
           } catch (Exception e) {
-            AndroidHelper.toast_long(c, c.getString(R.string.error) + ": " + e.getMessage());
+            UIHelper.toast_long(mContext, mContext.getString(R.string.error) + ": " + e.getMessage());
             Log.e(getClass().getSimpleName(), "Error: " + e.getMessage(), e);
           }
         }
@@ -188,12 +178,12 @@ public class DropboxImportExport implements DropboxListener{
 
   @Override
   public void onDroptboxListFoderError(Exception e) {
-    if(dialog != null)
-      dialog.dismiss();
+    if(mDialog != null)
+      mDialog.dismiss();
     Log.e(getClass().getSimpleName(), "Failed to get the file list.", e);
-    AndroidHelper.toast(c, c.getString(R.string.error_dropbox_list_directory));
-    if(dropboxDownloaded != null)
-      dropboxDownloaded.dropboxDownloaded(true);
+    UIHelper.toast(mContext, mContext.getString(R.string.error_dropbox_list_directory));
+    if(mDropboxDownloaded != null)
+      mDropboxDownloaded.dropboxDownloaded(true);
   }
 
   public static void loadDb(final Context c, File file) throws Exception{
@@ -202,19 +192,20 @@ public class DropboxImportExport implements DropboxListener{
     app.getDaysFactory().reload(app.getSql());
     app.getProfilesFactory().reload(app.getSql());
     app.getPublicHolidaysFactory().reload(app.getSql());
-    AndroidHelper.toast(c, c.getString(R.string.import_success));
+    UIHelper.toast(c, c.getString(R.string.import_success));
     AndroidHelper.restartApplication(c, -1);
   }
 
-  public static <T, V> void computeAndLoad(final Context c, final List<T> list, AndroidHelper.AlertDialogListListener<V> yes) {
+  public static <T, V> void computeAndLoad(final Context c, final List<T> list, UIHelper.AlertDialogListListener<V> yes) {
     List<String> files = compute(list);
     if(files.isEmpty())
-      AndroidHelper.toast_long(c, c.getString(R.string.error_no_files));
+      UIHelper.toast_long(c, c.getString(R.string.error_no_files));
     else {
-      AndroidHelper.showAlertDialog(c, R.string.box_select_db_file, files, yes);
+      UIHelper.showAlertDialog(c, R.string.box_select_db_file, files, yes);
     }
   }
 
+  @SuppressWarnings("unchecked")
   private static <T, V> List<V> compute(final List<T> list) {
     List<V> files = new ArrayList<>();
     for(T t : list) {
